backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
cryptography==42.0.7
python-dotenv==1.0.1
httpx==0.27.0
afip.py==1.2.0
